[+] add file rotation and compression to `JSONWriter`
JSONWriter opens or creates the `--json-storage-file` on first Write.
If the file exists and is less than 100Mb, JSONWriter will
open and append to that file. If the file exists and its size
is >= 100Mb, the file is renamed by putting the current time
in a timestamp in the name immediately before the file's extension.
A new file is then created using original filename.

Whenever a write would cause the current file exceed 100Mb, the current
file is closed, renamed, and a new file created with the original
name. Thus, the `--json-storage-file` value is always the "current" file.

Backups use the file name given, in the form `name-timestamp.ext` where
name is the filename without the extension, timestamp is the time at
which the file was rotated formatted with the `time.Time` format of
`2006-01-02T15-04-05.000` and the extension is the original extension.
For example, if `--json-storage-file` value is `/var/pw3/metrics.json`,
a backup created at 6:30pm on Nov 11 2024 would use the filename
`/var/pw3/metrics-2024-11-04T18-30-00.000.json`