[*] use `Metric` term for definition and `Measurement` for data gathered
Everywhere in code where we are talking about defining what and how to
monitor, we call it `Metric`. Everywhere in code where we are working
with the real data obtained from sources, we call that `Measurement`.

For example, "table_bloat" is a **metric** showing how many bloated
space a table occupies. And `"bloat_percent":2,"table_name":"foo",
"epoch_ns":1705078795154269000,"real_dbname":"pgwatch3"` is a
**measurement**.