[-] fix `QueryMeasurements` nil pointer dereference, fixes #766 (#767)
* keep existing connection in SourceConns.SyncFromReader()

Since `SourceConn` passed to reapMetricMeasurements() it's crucial
to keep the existing source with connection alive, so the reaper
keeps it's work. And there is no need to close the connection because
it might be used right now in some metric fetcher. Let Reaper to cancel
all disabled monitored sources in the main loop of Reap() function.

* fix tests