[-] add additional checks to `LogParse()`
Fix CPU consumption introduced in #fb33d09d. Add additional timer check.
Add `pprof` profile to demo Docker image. Add additional check if
pgwatch and monitored Postgres are running on the same host. Do not
start `LogParse` if pgwatch is unable to read logs and if
`--direct-os-stats` is not specified.