Make the next file chosing logic more precise.
Imagine a secnario where a huge load on the database
caused the current log file to get full and a new one was written to full
and a third one is written to half in very little time,
then all will have the same last modified time so
we must use `>=` also we want to traverse them
in order so we will add an additional sort by name.

The order by name is based on the assumption that
log files name will be monotonically increasing.

If the file names are cyclic e.g. by hours, then on a clock flip `24 => 1`
and the above scenario the intermediate file will get ignored (a bug).